requirements.txt
discord.py==2.3.2
wavelink==3.2.0
yt-dlp